+++ palace/drivers/basesolver.cpp
@@ -5,8 +5,10 @@
 
 #include <array>
 #include <complex>
+#include <numeric>
 #include <mfem.hpp>
 #include <nlohmann/json.hpp>
+#include "drivers/transientsolver.hpp"
 #include "fem/errorindicator.hpp"
 #include "fem/fespace.hpp"
 #include "linalg/ksp.hpp"
@@ -14,7 +16,9 @@
 #include "models/postoperator.hpp"
 #include "models/surfacepostoperator.hpp"
 #include "utils/communication.hpp"
+#include "utils/dorfler.hpp"
 #include "utils/filesystem.hpp"
+#include "utils/geodata.hpp"
 #include "utils/iodata.hpp"
 #include "utils/timer.hpp"
 
@@ -29,6 +33,38 @@ namespace
 std::string GetPostDir(const std::string &output)
 {
   return (output.length() > 0 && output.back() != '/') ? output + '/' : output;
+}
+
+std::string GetIterationPostDir(const std::string &output, int step, int width)
+{
+  return fmt::format("{}adapt{:0{}d}/", output, step, width);
+}
+
+void SaveIteration(MPI_Comm comm, const std::string &output, int step, int width)
+{
+  namespace fs = std::filesystem;
+  BlockTimer bt(Timer::IO);
+  Mpi::Barrier(comm);  // Wait for all processes to write postprocessing files
+  if (Mpi::Root(comm))
+  {
+    // Create a subfolder for the results of this adaptation.
+    const std::string step_output = GetIterationPostDir(output, step, width);
+    if (!fs::exists(step_output))
+    {
+      fs::create_directories(step_output);
+    }
+    constexpr auto options =
+        fs::copy_options::recursive | fs::copy_options::overwrite_existing;
+    for (const auto &f : fs::directory_iterator(output))
+    {
+      if (f.path().filename().string().rfind("adapt") == 0)
+      {
+        continue;
+      }
+      fs::copy(f, step_output + f.path().filename().string(), options);
+    }
+  }
+  Mpi::Barrier(comm);
 }
 
 json LoadMetadata(const std::string &post_dir)
@@ -51,6 +87,21 @@ void WriteMetadata(const std::string &post_dir, const json &meta)
     MFEM_ABORT("Unable to open metadata file \"" << path << "\"!");
   }
   fo << meta.dump(2) << '\n';
+}
+
+// Returns an array of indices corresponding to marked elements.
+mfem::Array<int> MarkedElements(const Vector &e, double threshold)
+{
+  mfem::Array<int> ind;
+  ind.Reserve(e.Size());
+  for (int i = 0; i < e.Size(); i++)
+  {
+    if (e[i] >= threshold)
+    {
+      ind.Append(i);
+    }
+  }
+  return ind;
 }
 
 }  // namespace
@@ -83,6 +134,111 @@ BaseSolver::BaseSolver(const IoData &iodata, bool root, int size, int num_thread
     }
     WriteMetadata(post_dir, meta);
   }
+}
+
+void BaseSolver::SolveEstimateMarkRefine(
+    std::vector<std::unique_ptr<mfem::ParMesh>> &mesh) const
+{
+  const auto &refinement = iodata.model.refinement;
+  const bool use_amr = [&]()
+  {
+    if (dynamic_cast<const TransientSolver *>(this) != nullptr)
+    {
+      Mpi::Warning("AMR is not currently supported for transient simulations!\n");
+      return false;
+    }
+    return refinement.max_it > 0;
+  }();
+  if (use_amr && mesh.size() > 1)
+  {
+    Mpi::Print("\nFlattening mesh sequence:\n AMR will start from the final mesh in "
+               "the sequence of a priori refinements\n");
+    mesh.erase(mesh.begin(), mesh.end() - 1);
+    constexpr bool refine = true, fix_orientation = true;
+    mesh.back()->Finalize(refine, fix_orientation);
+  }
+
+  // Perform initial solve and estimation.
+  MPI_Comm comm = mesh.back()->GetComm();
+  auto [indicators, ntdof] = Solve(mesh);
+  double err = indicators.Norml2(comm);
+
+  // Collection of all tests that might exhaust resources.
+  auto ExhaustedResources = [&refinement](auto it, auto ntdof)
+  {
+    bool ret = false;
+    // Run out of iterations.
+    ret |= (it >= refinement.max_it);
+    // Run out of DOFs if a limit was set.
+    ret |= (refinement.max_size >= 1 && ntdof > refinement.max_size);
+    return ret;
+  };
+
+  // Main AMR loop.
+  int it = 0;
+  while (!ExhaustedResources(it, ntdof) && err >= refinement.tol)
+  {
+    BlockTimer bt(Timer::ADAPTATION);
+    Mpi::Print("\nAdaptive mesh refinement (AMR) iteration {:d}:\n"
+               " Indicator norm = {:.3e}, size = {:d}\n"
+               " Maximum iterations = {:d}, tol. = {:.3e}{}\n",
+               ++it, err, ntdof, refinement.max_it, refinement.tol,
+               (refinement.max_size > 0
+                    ? ", maximum size = " + std::to_string(refinement.max_size)
+                    : ""));
+
+    // Optionally save off the previous solution.
+    if (refinement.save_adapt_iterations)
+    {
+      SaveIteration(comm, post_dir, it,
+                    1 + static_cast<int>(std::log10(refinement.max_it)));
+    }
+
+    // Mark.
+    const auto [threshold, marked_error] = utils::ComputeDorflerThreshold(
+        comm, indicators.Local(), refinement.update_fraction);
+    const auto marked_elements = MarkedElements(indicators.Local(), threshold);
+    const auto [glob_marked_elements, glob_elements] =
+        linalg::GlobalSize2(comm, marked_elements, indicators.Local());
+    Mpi::Print(
+        " Marked {:d}/{:d} elements for refinement ({:.2f}% of the error, θ = {:.2f})\n",
+        glob_marked_elements, glob_elements, 100 * marked_error,
+        refinement.update_fraction);
+
+    // Refine.
+    const auto initial_elem_count = mesh.back()->GetGlobalNE();
+    mesh.back()->GeneralRefinement(marked_elements, -1, refinement.max_nc_levels);
+    const auto final_elem_count = mesh.back()->GetGlobalNE();
+    Mpi::Print(" Mesh refinement added {:d} elements (initial: {}, final: {})\n",
+               final_elem_count - initial_elem_count, initial_elem_count, final_elem_count);
+
+    // Optionally rebalance and write the adapted mesh to file.
+    const auto ratio_pre =
+        mesh::RebalanceMesh(iodata, mesh.back(), refinement.maximum_imbalance);
+    if (ratio_pre > refinement.maximum_imbalance)
+    {
+      int min_elem, max_elem;
+      min_elem = max_elem = mesh.back()->GetNE();
+      Mpi::GlobalMin(1, &min_elem, comm);
+      Mpi::GlobalMax(1, &max_elem, comm);
+      const auto ratio_post = double(max_elem) / min_elem;
+      Mpi::Print(" Rebalanced mesh: Ratio {:.3f} exceeded maximum allowed value {:.3f} "
+                 "(new ratio = {:.3f})\n",
+                 ratio_pre, refinement.maximum_imbalance, ratio_post);
+    }
+
+    // Solve + estimate.
+    Mpi::Print("\nProceeding with solve/estimate iteration {}...\n", 1 + it);
+    std::tie(indicators, ntdof) = Solve(mesh);
+    err = indicators.Norml2(comm);
+  }
+  Mpi::Print("\nCompleted {:d} iteration{} of adaptive mesh refinement (AMR):\n"
+             " Indicator norm = {:.3e}, size = {:d}\n"
+             " Maximum iterations = {:d}, tol. = {:.3e}{}\n",
+             it, (it == 1 ? "" : "s"), err, ntdof, refinement.max_it, refinement.tol,
+             (refinement.max_size > 0
+                  ? ", maximum size = " + std::to_string(refinement.max_size)
+                  : ""));
 }
 
 void BaseSolver::SaveMetadata(const FiniteElementSpaceHierarchy &fespaces) const