+++ palace/drivers/drivensolver.hpp
@@ -60,10 +60,11 @@ private:
                               const WavePortOperator &wave_port_op, int step,
                               double omega) const;
 
+  std::pair<ErrorIndicator, long long int>
+  Solve(const std::vector<std::unique_ptr<mfem::ParMesh>> &mesh) const override;
+
 public:
   using BaseSolver::BaseSolver;
-
-  ErrorIndicator Solve(const std::vector<std::unique_ptr<mfem::ParMesh>> &mesh) const final;
 };
 
 }  // namespace palace