+++ palace/drivers/electrostaticsolver.hpp
@@ -42,11 +42,11 @@ private:
                             const mfem::DenseMatrix &C, const mfem::DenseMatrix &Cinv,
                             const mfem::DenseMatrix &Cm) const;
 
+  std::pair<ErrorIndicator, long long int>
+  Solve(const std::vector<std::unique_ptr<mfem::ParMesh>> &mesh) const override;
+
 public:
   using BaseSolver::BaseSolver;
-
-  ErrorIndicator
-  Solve(const std::vector<std::unique_ptr<mfem::ParMesh>> &mesh) const override;
 };
 
 }  // namespace palace