+++ palace/main.cpp
@@ -285,7 +285,7 @@ int main(int argc, char *argv[])
   mesh::RefineMesh(iodata, mesh);
 
   // Run the problem driver.
-  auto indicators = solver->Solve(mesh);
+  solver->SolveEstimateMarkRefine(mesh);
 
   // Print timing summary.
   BlockTimer::Print(world_comm);